changed doubles to floats

+++ functions/msfm2d_MOD.h
@@ -1,8 +1,8 @@
 #ifndef MSFM2D_MOD_H
 #define MSFM2D_MOD_H
 
-double* main_msfm(double* F, double* source_points, double* T, int* size_map, int* size_target);
-double* velocities_map(double* binary_map, int* size_map, int threshold, double safety_margin);
-void compute_gradient_2d_discrete(double* input_matrix, double* gradient_matrix, int* size_map);
+float* main_msfm(float* F, float* source_points, float* T, int* size_map, int* size_target);
+float* velocities_map(float* binary_map, int* size_map, int threshold, float safety_margin);
+void compute_gradient_2d_discrete(float* input_matrix, float* gradient_matrix, int* size_map);
 
 #endif
